Move back upgrade to configure to reduce PR diff size
