+++ src/cli/configure/upgrade/patches/7.3.1/addEmptyExports.test.ts
